👌 IMP: Move poilicy evaluation into evaluation module
